split options selectors into multiple files
closes #400
